+++ shell/shell.go
@@ -57,7 +57,7 @@ func RunShell(apiCtx *api.ApiCtx) error {
 	if(hasAllToolsForGetaInstalled){
 		shell.AddCmd(getaCmd(ctx))
 		shell.AddCmd(mgetaCmd(ctx))
-		shell.AddCmd(syncCmd(ctx))
+		shell.AddCmd(mirrorCmd(ctx))
 	} else {
 		log.Warning.Println(fmt.Sprintf("Commands geta, mgeta and sync are disabled" + 
 			" because the following tools are not installed: \n %v",